requirements.txt
python-dotenv>=1.0.1
requests>=2.32.0
Pillow>=10.3.0
openai>=1.35.0 